+++ src/app/modules/search-parameter/observation-test-value.component.ts
@@ -28,6 +28,8 @@ export class ObservationTestValueComponent
   extends BaseControlValueAccessor<ObservationTestValue>
   implements OnInit {
   @Input() datatype: string;
+  @Input() loincCodes: string[] = [];
+
   form = new FormGroup({
     testValuePrefix: new FormControl(''),
     testValueModifier: new FormControl(''),
@@ -38,7 +40,9 @@ export class ObservationTestValueComponent
   readonly typeDescriptions = {
     Quantity: {
       searchValPrefixes: [
-        ['=', 'eq'],
+        // See https://www.hl7.org/fhir/search.html#prefix
+        // if no prefix is present, the prefix 'eq' is assumed.
+        ['=', ''],
         ['not equal', 'ne'],
         ['>', 'gt'],
         ['<', 'lt'],